+++ app/javascript/engine/Map.ts
@@ -677,7 +677,17 @@ export default class Map {
   enemyAt(loc: Coordinate, player: number): boolean {
     const nation = player === 1 ? this.game?.playerOneNation : this.game?.playerTwoNation
     for (const c of this.countersAt(loc)) {
-      if (c.hasUnit && !c.unit.isWreck) {
+      if (c.hasUnit && !c.unit.isWreck && !c.unit.operated) {
+        if (c.unit.playerNation !== nation) { return true }
+      }
+    }
+    return false
+  }
+
+  vpEnemyAt(loc: Coordinate, player: number): boolean {
+    const nation = player === 1 ? this.game?.playerOneNation : this.game?.playerTwoNation
+    for (const c of this.countersAt(loc)) {
+      if (c.hasUnit && !c.unit.isWreck && !c.unit.operated && !c.unit.leader ) {
         if (c.unit.playerNation !== nation) { return true }
       }
     }